CHECKER/requirements.txt
pyfiglet==0.8.post1
termcolor==2.3.0
requests
dnspython
argparse
